[#185] move inject package into upper lever
