+++ aws-wildfly/src/main/java/sunstone/aws/impl/AwsWFIdentifiableSunstoneResourceUtils.java
@@ -5,7 +5,7 @@ import sunstone.annotation.WildFly;
 import org.wildfly.extras.creaper.core.online.OnlineManagementClient;
 import software.amazon.awssdk.services.ec2.model.Instance;
 import sunstone.annotation.OperatingMode;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 import sunstone.core.CreaperUtils;
 import sunstone.core.exceptions.SunstoneException;
 import sunstone.core.exceptions.UnsupportedSunstoneOperationException;
