+++ aws-wildfly/src/test/java/aws/cloudformation/archiveDeploy/ec2/suitetests/AwsEc2UndeployedSecondTest.java
@@ -10,7 +10,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import sunstone.annotation.Parameter;
 import sunstone.aws.annotation.WithAwsCfTemplate;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 
 import java.io.IOException;
