+++ aws-wildfly/src/test/java/aws/cloudformation/archiveDeploy/ec2Domain/suitetests/AwsDomainEc2DeployFirstTest.java
@@ -16,7 +16,7 @@ import org.junit.jupiter.api.Test;
 import sunstone.annotation.Deployment;
 import sunstone.annotation.Parameter;
 import sunstone.aws.annotation.WithAwsCfTemplate;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 
 import java.io.IOException;
