+++ aws-wildfly/src/test/java/aws/cloudformation/archiveDeploy/ec2Domain/suitetests/AwsDomainEc2UndeployedSecondTest.java
@@ -12,7 +12,7 @@ import org.assertj.core.api.Assertions;
 import org.junit.jupiter.api.Test;
 import sunstone.annotation.Parameter;
 import sunstone.aws.annotation.WithAwsCfTemplate;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 
 import java.io.IOException;
