+++ aws/src/main/java/sunstone/aws/impl/AwsSunstoneResourceInjector.java
@@ -9,7 +9,7 @@ import org.junit.jupiter.api.extension.ExtensionContext;
 import sunstone.core.SunstoneConfig;
 import software.amazon.awssdk.services.ec2.Ec2Client;
 import software.amazon.awssdk.services.s3.S3Client;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 import sunstone.core.api.SunstoneResourceInjector;
 import sunstone.core.exceptions.IllegalArgumentSunstoneException;
 import sunstone.core.exceptions.SunstoneException;
