+++ aws/src/test/java/aws/cloudformation/di/suitetests/AwsHostnameTests.java
@@ -6,7 +6,7 @@ import org.junit.jupiter.api.BeforeAll;
 import org.junit.jupiter.api.Test;
 import sunstone.annotation.Parameter;
 import sunstone.aws.annotation.WithAwsCfTemplate;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 
 import static aws.cloudformation.AwsTestConstants.instanceName;
 import static aws.cloudformation.AwsTestConstants.region;
