+++ azure-wildfly/src/main/java/sunstone/azure/impl/AzureWFIdentifiableSunstoneResourceUtils.java
@@ -6,7 +6,7 @@ import com.azure.resourcemanager.appservice.models.WebApp;
 import com.azure.resourcemanager.compute.models.VirtualMachine;
 import org.wildfly.extras.creaper.core.online.OnlineManagementClient;
 import sunstone.annotation.OperatingMode;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 import sunstone.core.CreaperUtils;
 import sunstone.core.exceptions.SunstoneException;
 import sunstone.core.exceptions.UnsupportedSunstoneOperationException;
