+++ azure-wildfly/src/test/java/sunstone/azure/armTemplates/archiveDeploy/vm/suitetests/AzureVmDeployFirstTest.java
@@ -14,7 +14,7 @@ import org.junit.jupiter.api.Test;
 import sunstone.annotation.Parameter;
 import sunstone.annotation.Deployment;
 import sunstone.azure.annotation.WithAzureArmTemplate;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 
 import java.io.IOException;
