+++ azure-wildfly/src/test/java/sunstone/azure/armTemplates/archiveDeploy/webapp/suitetests/AzureWebAppDeployFirstTest.java
@@ -15,7 +15,7 @@ import org.junit.jupiter.api.Test;
 import sunstone.annotation.Deployment;
 import sunstone.annotation.Parameter;
 import sunstone.azure.annotation.WithAzureArmTemplate;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 
 import java.io.IOException;
