+++ azure/src/main/java/sunstone/azure/impl/AzureIdentifiableSunstoneResourceUtils.java
@@ -3,7 +3,7 @@ package sunstone.azure.impl;
 
 import com.azure.resourcemanager.appservice.models.WebApp;
 import com.azure.resourcemanager.compute.models.VirtualMachine;
-import sunstone.annotation.inject.Hostname;
+import sunstone.inject.Hostname;
 import sunstone.core.exceptions.SunstoneException;
 import sunstone.core.exceptions.UnsupportedSunstoneOperationException;
